feat(ui): add Sentry.io for error monitoring. Minor refactors
Added Sentry.io for monitoring and tracing of (un)caught/(un)handled errors occuring
on the frontend.
Minor adjustments to some components where refactoring or code splitting were needed,
implement useQuery instead of useEffects.
Fixed bug in CreditsTable where if you selecteda (just one) course and then removed it instantly,
the master check table still rendered its credits etc. while removed (stale data).
