+++ StudyPlannerUI/.gitignore
@@ -42,3 +42,5 @@ dist-ssr
 **/.vs/*
 **/obj/*
 **/bin/*
+# Sentry Config File
+.env.sentry-build-plugin
