+++ StudyPlannerUI/package.json
@@ -8,6 +8,7 @@
     "build": "rm -rf build && tsc && vite build",
     "test": "vitest",
     "preview": "vite preview",
+    "serve": "pnpm build && serve -s build -l 5173",
     "check:unused-code": "ts-prune | grep -v '(used in module)'",
     "check:dupe-code": "jscpd src",
     "prettier:check": "prettier --check src",
@@ -17,6 +18,8 @@
     "prepare": "cd ../ && husky install ./StudyPlannerUI/.husky"
   },
   "dependencies": {
+    "@sentry/react": "^7.99.0",
+    "@sentry/vite-plugin": "^2.11.0",
     "@tanstack/react-query": "^5.18.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -64,6 +67,7 @@
     "prettier": "^2.8.8",
     "rollup-plugin-analyzer": "^4.0.0",
     "rollup-plugin-visualizer": "^5.9.0",
+    "serve": "^14.2.1",
     "stylelint": "15.6.0",
     "stylelint-config-prettier": "9.0.5",
     "stylelint-config-standard": "33.0.0",
